Going down the list of packages removed from environment.yaml and replacing with soft_import call.  Just finished matplotlib (with some extras).
